Augmented StereoSubscriber
StereoSubscriber now holds references to the originally received
messages for the subclasses to access. This is necessary in a few cases.

StereoSubscriber has implementations of Subscribers, for Images as well
as Calibration data. These implementations call pure virtual
implementations for subclasses to override. Added the one for
calibration data, previously there was only one for image data.

Added empty implementation of pure virtual method in stereo_view

	modified:   include/cvt_ros/StereoSubscriber.h
	modified:   src/lib/StereoSubscriber.cpp
	modified:   src/stereo_view.cpp